requirements.txt
requests
python-dotenv
flask
dashscope
sentence-transformers
faiss-cpu
